+++ CHANGELOG.md
@@ -5,6 +5,7 @@ that call deleted endpoints is not considered a breaking change.
 
 ## 3.18.0 (Unreleased)
 
+* Add `get_order_statuses` for the new `order/orderstatuses` route
 * Support another weird HTML formatting for 500 errors
 * Allow to import `BBShippingError` and `BBWarehouseSplitError` directly from `bigbuy`